Keep Request Insights data out of DevTools traces
